+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/HistoryDao.java
@@ -3,10 +3,10 @@ package com.senla.haltvinizki.dao;
 import com.senla.haltvinizki.entity.History;
 
 
-public interface HistoryDao extends GenericDao<History, Integer> {
-    History getHistoryWithProduct(int id);
+public interface HistoryDao extends GenericDao<History, Long> {
+    History getHistoryWithProduct(Long id);
 
-    History getHistoryWithCustomer(int id);
+    History getHistoryWithCustomer(Long id);
 
 
 }
