+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/ProductConfigurationDao.java
@@ -2,7 +2,7 @@ package com.senla.haltvinizki.dao;
 
 import com.senla.haltvinizki.entity.ProductConfiguration;
 
-public interface ProductConfigurationDao extends GenericDao<ProductConfiguration, Integer> {
-    ProductConfiguration getProductConfigWithProduct(int id);
+public interface ProductConfigurationDao extends GenericDao<ProductConfiguration, Long> {
+    ProductConfiguration getProductConfigWithProduct(Long id);
 
 }
