+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/ProductDao.java
@@ -4,12 +4,12 @@ import com.senla.haltvinizki.entity.Product;
 
 import java.util.List;
 
-public interface ProductDao extends GenericDao<Product,Integer>{
+public interface ProductDao extends GenericDao<Product,Long>{
     Product getMostExpensiveProduct();
 
     List<Product> getActiveProducts();
 
-    Product getProductWithUser(int id);
+    Product getProductWithUser(Long id);
 
-    Product getProductWithCategory(int id);
+    Product getProductWithCategory(Long id);
 }
