+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/RoleDao.java
@@ -2,8 +2,8 @@ package com.senla.haltvinizki.dao;
 
 import com.senla.haltvinizki.entity.Role;
 
-public interface RoleDao extends GenericDao<Role, Integer> {
+public interface RoleDao extends GenericDao<Role, Long> {
 
-    Role getRoleWithUsers(int id);
+    Role getRoleWithUsers(Long id);
 
 }
